Stop truncating code lengths to 16 bits in minimum-weight
The packed matrix and INFO structs counted with unsigned short, and the
enumeration loops accumulated codeword weights in short.  A code of
length 65539 therefore recorded 65536 generator matrices as 0 and a
block length of 65539 as 3, and the program read past its allocations
and crashed instead of answering 2:

    3 x 65539 over GF(2), rows [1]*65536 + e_i, minimum distance 2
    before: "There are 0 generator matrices" then SIGSEGV
    after:  "Minimum weight: 2"

Widen the lengths, counts and weights to unsigned int, which covers any
code GAP can hold.  The GF(3) arithmetic tables stay 16-bit: they hold
field elements, not lengths.

generator_matrix() trusted the file it was handed.  It read %d into
unsigned int fields, ignored how many of the three header values it got,
sized allocations by multiplying an attacker-chosen column count by
sizeof(unsigned int), never checked a malloc, and its caller ignored its
result.  It now validates the header, multiplies in size_t and rejects
what would wrap, reports allocation failure, and stops on a short read.

make check-programs covers both: the GAP code never builds a code long
enough to reach the first, or a malformed file to reach the second.

Fixes https://github.com/gap-packages/guava/issues/46
Fixes https://github.com/gap-packages/guava/issues/57

+++ src/ctjhai/minimum-weight-gf2.c
@@ -73,7 +73,7 @@ int mindist_gf2(MATRIX G, mod_t m_mod, int lower_bound) {
 	 */
 	j = (G.cols - G.rows) + 1;	/* maximum possible number of matrices */
 	packedG.mat = (packed_t ***)malloc(j * sizeof(packed_t **));
-	packedG.rank= (unsigned short *)malloc(j * sizeof(unsigned short));
+	packedG.rank= (unsigned int *)malloc(j * sizeof(unsigned int));
 	packedG.dimension = G.rows;
 	packedG.block_length = G.cols;
 	packedG.nMatrices = packedG.nFullRankMatrices = 0;
@@ -190,7 +190,7 @@ int cyclic_mindist_gf2(MATRIX G, mod_t m_mod, int lower_bound) {
 	 */
 	j = 1;	/* only one generator matrix */
 	packedG.mat = (packed_t ***)malloc(j * sizeof(packed_t **));
-	packedG.rank= (unsigned short *)malloc(j * sizeof(unsigned short));
+	packedG.rank= (unsigned int *)malloc(j * sizeof(unsigned int));
 	packedG.dimension = G.rows;
 	packedG.block_length = G.cols;
 	packedG.nMatrices = j; 
@@ -317,10 +317,10 @@ void __mindist_gf2(PACKED_MATRIX_GF2 *G, INFO *info) {
 		} while (l);\
 	}
 	double steps;
-	short m, iw, top, In, Out, nMat;
-	register short i, j, l, w;
+	int m, iw, top, In, Out, nMat;
+	register int i, j, l, w;
 	register packed_t **c;
-	short *v, *s;
+	int *v, *s;
 
 	w = top = 0;
 	nMat = info->matrices_req;
@@ -336,8 +336,8 @@ void __mindist_gf2(PACKED_MATRIX_GF2 *G, INFO *info) {
 		steps = 0;
 		
 		m = iw - 1;
-		v = (short *) malloc((iw+2)*sizeof(short));
-		s = (short *) malloc((iw+2)*sizeof(short));
+		v = (int *) malloc((iw+2)* sizeof(int));
+		s = (int *) malloc((iw+2)* sizeof(int));
 		
 		if ( !(iw & 1) ) {
 			v[iw + 1] = G->dimension; v[iw] = iw - 1;
@@ -453,10 +453,10 @@ void __cyclic_mindist_gf2(PACKED_MATRIX_GF2 *G, INFO *info) {
 		}\
 	}
 	double steps;
-	short m, iw, top, In, Out;
-	register short i, j, w;
+	int m, iw, top, In, Out;
+	register int i, j, w;
 	register packed_t *c;
-	short *v, *s;
+	int *v, *s;
 
 	w = top = 0;
 	c = (packed_t *)malloc(G->nBlocks * sizeof(packed_t)); 
@@ -467,8 +467,8 @@ void __cyclic_mindist_gf2(PACKED_MATRIX_GF2 *G, INFO *info) {
 		steps = 0;
 		
 		m = iw - 1;
-		v = (short *) malloc((iw+2)*sizeof(short));
-		s = (short *) malloc((iw+2)*sizeof(short));
+		v = (int *) malloc((iw+2)* sizeof(int));
+		s = (int *) malloc((iw+2)* sizeof(int));
 		
 		if ( !(iw & 1) ) {
 			v[iw + 1] = G->dimension; v[iw] = iw - 1;
@@ -553,7 +553,7 @@ lower_bound_met:
 /* current progress of enumeration.                                 */
 void update_info_gf2(PACKED_MATRIX_GF2 *G, INFO *info) {
 	bool end;
-	short i, j, l, d, w;
+	int i, j, l, d, w;
 	
 	/* Estimate the number of matrices required to complete enumeration */
 	info->matrices_req = 0;
@@ -633,7 +633,7 @@ void update_info_gf2(PACKED_MATRIX_GF2 *G, INFO *info) {
 /* Update INFO structure, which contains various information on the */
 /* current progress of enumeration.                                 */
 void cyclic_update_info_gf2(PACKED_MATRIX_GF2 *G, INFO *info) {
-	short i, d, w;
+	int i, d, w;
 	
 	/* Estimate the minimum distance lower bound after current enumeration */
 	info->lower_bound = (info->info_weight_completed + 1) * G->nFullRankMatrices;
